Moved repeated functions into a common helper object
