+++ hudi-spark-datasource/hudi-spark2/src/main/scala/org/apache/spark/sql/execution/datasources/parquet/Spark24HoodieParquetFileFormat.scala
@@ -25,7 +25,6 @@ import org.apache.hadoop.mapreduce.{JobID, TaskAttemptID, TaskID, TaskType}
 import org.apache.parquet.filter2.compat.FilterCompat
 import org.apache.parquet.filter2.predicate.FilterApi
 import org.apache.parquet.format.converter.ParquetMetadataConverter.SKIP_ROW_GROUPS
-import org.apache.parquet.hadoop.metadata.FileMetaData
 import org.apache.parquet.hadoop.{ParquetFileReader, ParquetInputFormat, ParquetRecordReader}
 import org.apache.spark.TaskContext
 import org.apache.spark.sql.SparkSession
@@ -37,7 +36,7 @@ import org.apache.spark.sql.catalyst.util.DateTimeUtils
 import org.apache.spark.sql.execution.datasources.{PartitionedFile, RecordReaderIterator}
 import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.sources.Filter
-import org.apache.spark.sql.types.{ArrayType, AtomicType, DataType, MapType, StructField, StructType}
+import org.apache.spark.sql.types.{AtomicType, StructField, StructType}
 import org.apache.spark.util.SerializableConfiguration
 
 import java.net.URI
@@ -163,7 +162,7 @@ class Spark24HoodieParquetFileFormat(private val shouldAppendPartitionValues: Bo
 
       // Clone new conf
       val hadoopAttemptConf = new Configuration(broadcastedHadoopConf.value.value)
-      val (implicitTypeChangeInfos, sparkRequestSchema) = buildImplicitSchemaChangeInfo(hadoopAttemptConf, footerFileMetaData, requiredSchema)
+      val (implicitTypeChangeInfos, sparkRequestSchema) = HoodieParquetFileFormatHelper.buildImplicitSchemaChangeInfo(hadoopAttemptConf, footerFileMetaData, requiredSchema)
 
       if (!implicitTypeChangeInfos.isEmpty) {
         hadoopAttemptConf.set(ParquetReadSupport.SPARK_ROW_REQUESTED_SCHEMA, sparkRequestSchema.json)
@@ -263,50 +262,4 @@ class Spark24HoodieParquetFileFormat(private val shouldAppendPartitionValues: Bo
     }
   }
 
-  private def buildImplicitSchemaChangeInfo(hadoopConf: Configuration,
-                                            parquetFileMetaData: FileMetaData,
-                                            requiredSchema: StructType): (java.util.Map[Integer, org.apache.hudi.common.util.collection.Pair[DataType, DataType]], StructType) = {
-    val implicitTypeChangeInfo: java.util.Map[Integer, org.apache.hudi.common.util.collection.Pair[DataType, DataType]] = new java.util.HashMap()
-    val convert = new ParquetToSparkSchemaConverter(hadoopConf)
-    val fileStruct = convert.convert(parquetFileMetaData.getSchema)
-    val fileStructMap = fileStruct.fields.map(f => (f.name, f.dataType)).toMap
-    val sparkRequestStructFields = requiredSchema.map(f => {
-      val requiredType = f.dataType
-      if (fileStructMap.contains(f.name) && !isDataTypeEqual(requiredType, fileStructMap(f.name))) {
-        implicitTypeChangeInfo.put(new Integer(requiredSchema.fieldIndex(f.name)), org.apache.hudi.common.util.collection.Pair.of(requiredType, fileStructMap(f.name)))
-        StructField(f.name, fileStructMap(f.name), f.nullable)
-      } else {
-        f
-      }
-    })
-    (implicitTypeChangeInfo, StructType(sparkRequestStructFields))
-  }
-
-  private def isDataTypeEqual(requiredType: DataType, fileType: DataType): Boolean = (requiredType, fileType) match {
-    case (requiredType, fileType) if requiredType == fileType => true
-
-    case (ArrayType(rt, _), ArrayType(ft, _)) =>
-      // Do not care about nullability as schema evolution require fields to be nullable
-      isDataTypeEqual(rt, ft)
-
-    case (MapType(requiredKey, requiredValue, _), MapType(fileKey, fileValue, _)) =>
-      // Likewise, do not care about nullability as schema evolution require fields to be nullable
-      isDataTypeEqual(requiredKey, fileKey) && isDataTypeEqual(requiredValue, fileValue)
-
-    case (StructType(requiredFields), StructType(fileFields)) =>
-      // Find fields that are in requiredFields and fileFields as they might not be the same during add column + change column operations
-      val commonFieldNames = requiredFields.map(_.name) intersect fileFields.map(_.name)
-
-      // Need to match by name instead of StructField as name will stay the same whilst type may change
-      val fileFilteredFields = fileFields.filter(f => commonFieldNames.contains(f.name)).sortWith(_.name < _.name)
-      val requiredFilteredFields = requiredFields.filter(f => commonFieldNames.contains(f.name)).sortWith(_.name < _.name)
-
-      // Sorting ensures that the same field names are being compared for type differences
-      requiredFilteredFields.zip(fileFilteredFields).forall {
-        case (requiredField, fileFilteredField) =>
-          isDataTypeEqual(requiredField.dataType, fileFilteredField.dataType)
-      }
-
-    case _ => false
-  }
 }
